# openapi.yaml
openapi: '3.0.3'
info:
  title: Multi-Tenant Project Management API
  description: |
    RESTful API for a multi-tenant project management system with tenant isolation via API keys,
    tamper-evident audit trails, sliding window rate limiting, and queue-based email engine.
  version: '1.0.0'
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: 'API key in format: vz_xxxx'
    InternalKey:
      type: apiKey
      in: header
      name: x-internal-key
...
paths:
...
  /auth/register:
    post:
      summary: Register new tenant with owner
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantName, tenantSlug, ownerEmail, ownerName, password]
              properties:
                tenantName: { type: string }
                tenantSlug: { type: string }
                ownerEmail: { type: string }
                ownerName: { type: string }
                password: { type: string }
      responses:
        '201':
          description: Tenant and owner created
        '409':
          description: Tenant slug already exists
...
  /auth/api-keys/{keyId}/rotate:
    post:
      summary: Rotate API key (15-min grace period)
      tags: [Authentication]
      parameters:
        - name: keyId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Key rotated successfully
...
  /projects/{projectId}/tasks:
    get:
      summary: List tasks in project
      tags: [Tasks]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [todo, in-progress, completed] }
        - name: assignedToId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated list of tasks
    post:
      summary: Create task in project
      tags: [Tasks]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                status: { type: string }
      responses:
        '201':
          description: Task created
  /projects/{projectId}/tasks/{taskId}:
    get:
      summary: Get task details
      tags: [Tasks]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Task details
    put:
      summary: Update task
      tags: [Tasks]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Task updated
    delete:
      summary: Delete task
      tags: [Tasks]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Task deleted
...